pyproject.toml
[tool.poetry]
name = "teseo-advisor-webhook"
version = "0.1.0"
description = ""
authors = ["Gabriele Armento <gabriele.armento@gmail.com>"]
readme = "README.md"
packages = [{include = "teseo_advisor_webhook"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
python-decouple = "^3.8"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
